fix MSVC warnings

+++ AppendClip/AppendClip.cpp
@@ -643,7 +643,7 @@ AppendClipPlugin::setupAndProcess(ImageBlenderBase &processor,
     assert(0 < alpha0 && alpha0 <= 1 && 0 <= alpha1 && alpha1 < 1);
     assert( toImg.get() || (alpha1 == 0) );
     assert( fromImg.get() );
-    processor.setBlend(1. - alpha0);
+    processor.setBlend(static_cast<float>(1. - alpha0));
 
     // Call the base class process member, this will call the derived templated process code
     processor.process();
